# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wren (wren-external)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/roadmaps": {
      "post": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Create Roadmap",
        "operationId": "create_roadmap_roadmaps_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoadmapInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}": {
      "get": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Get Roadmap",
        "operationId": "get_roadmap_roadmaps__roadmap_id__get",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roadmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Patch Roadmap",
        "operationId": "patch_roadmap_roadmaps__roadmap_id__patch",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Replace Roadmap",
        "operationId": "replace_roadmap_roadmaps__roadmap_id__put",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoadmapInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapReplaced"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Delete Roadmap",
        "operationId": "delete_roadmap_roadmaps__roadmap_id__delete",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}/overview": {
      "get": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Get Overview",
        "operationId": "get_overview_roadmaps__roadmap_id__overview_get",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ResponseFormat",
              "default": "concise"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}/nodes/{subsection_id}": {
      "get": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Get Node",
        "operationId": "get_node_roadmaps__roadmap_id__nodes__subsection_id__get",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          },
          {
            "name": "subsection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subsection Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ResponseFormat",
              "default": "concise"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}/sections/{section_id}": {
      "get": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Get Section",
        "operationId": "get_section_roadmaps__roadmap_id__sections__section_id__get",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          },
          {
            "name": "section_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Section Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SectionInclude",
              "default": "both"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}/search": {
      "get": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Search Roadmap",
        "operationId": "search_roadmap_roadmaps__roadmap_id__search_get",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchHit"
                  },
                  "title": "Response Search Roadmap Roadmaps  Roadmap Id  Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}:validate": {
      "post": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Validate Roadmap",
        "operationId": "validate_roadmap_roadmaps__roadmap_id__validate_post",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}:publish": {
      "post": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Publish Roadmap",
        "operationId": "publish_roadmap_roadmaps__roadmap_id__publish_post",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roadmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}:fork": {
      "post": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Fork Roadmap",
        "operationId": "fork_roadmap_roadmaps__roadmap_id__fork_post",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roadmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}/metadata": {
      "patch": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Edit Roadmap Metadata",
        "operationId": "edit_roadmap_metadata_roadmaps__roadmap_id__metadata_patch",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetadataEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roadmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}/visibility": {
      "put": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Set Roadmap Visibility",
        "operationId": "set_roadmap_visibility_roadmaps__roadmap_id__visibility_put",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisibilityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roadmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/roadmaps/{roadmap_id}:archive": {
      "post": {
        "tags": [
          "roadmaps"
        ],
        "summary": "Archive Roadmap",
        "operationId": "archive_roadmap_roadmaps__roadmap_id__archive_post",
        "parameters": [
          {
            "name": "roadmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Roadmap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roadmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatchResult": {
        "properties": {
          "roadmap_id": {
            "type": "string",
            "title": "Roadmap Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "changed_nodes": {
            "items": {
              "$ref": "#/components/schemas/ChangedNode"
            },
            "type": "array",
            "title": "Changed Nodes"
          },
          "remap": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Remap"
          }
        },
        "type": "object",
        "required": [
          "roadmap_id",
          "revision"
        ],
        "title": "PatchResult",
        "description": "The ``PATCH /roadmaps/{id}`` body: the post-batch ``revision``, the changed\nnodes, and the ``proposed_id -> minted_id`` remap for any de-duped ``add_*``\nproposal. ``remap`` is empty when nothing was de-duped."
      },
...
      "ResponseFormat": {
        "type": "string",
        "enum": [
          "concise",
          "detailed"
        ],
        "title": "ResponseFormat",
        "description": "The ``concise | detailed`` switch on the read tools.\n\nConcise is roughly one-third the tokens and still carries the follow-up IDs;\ndetailed adds the explanatory free-text (the node ``description``)."
      },
...
      "RoadmapCreated": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "subject_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subject Tags"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility",
            "default": "private"
          },
          "status": {
            "$ref": "#/components/schemas/RoadmapStatus",
            "default": "draft"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 1
          },
          "sections": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "object",
            "title": "Sections"
          },
          "section_order": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Section Order"
          },
          "suggested_path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Path"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "remap": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Remap"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "RoadmapCreated",
        "description": "The ``POST /roadmaps`` body: the full minted roadmap plus a ``remap`` of\nevery de-duped ``proposed_id -> minted_id`` so the author can reconcile the\nreferences it sent. ``remap`` is empty when no proposed ID\nhad to be changed."
      },
      "RoadmapInput": {
        "properties": {
          "proposed_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "subject_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subject Tags"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility",
            "default": "private"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionInput"
            },
            "type": "array",
            "title": "Sections"
          },
          "suggested_path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Path"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "RoadmapInput",
        "description": "The ``create_roadmap_draft`` / ``replace_roadmap_draft`` payload."
      },
...
    }
  }
}